third_party/blink/web_tests/platform/generic/external/wpt/svg/geometry/parsing/rx-invalid-expected.txt

This is a testharness.js-based test.
FAIL e.style['rx'] = "10" should not set the property value assert_equals: expected "" but got "10"
PASS e.style['rx'] = "none" should not set the property value
PASS e.style['rx'] = "10px 20px" should not set the property value
PASS e.style['rx'] = "-1px" should not set the property value
Harness: the test ran to completion.
